+++ src/binary.rs
@@ -5,16 +5,19 @@
 //! This module defines the [`Binary`] trait for programs that may be traced as
 //! well as a number of types that may serve as a [`Binary`].
 
+pub mod basic;
+pub mod combinators;
 pub mod error;
 
 #[cfg(feature = "elf")]
 pub mod elf;
 
-use core::borrow::BorrowMut;
+pub use basic::{from_fn, Empty};
+pub use combinators::Multi;
 
 use crate::instruction::Instruction;
 
-use error::{MaybeMiss, Miss};
+use error::Miss;
 
 /// A binary of some sort that contains [`Instruction`]s
 pub trait Binary {
@@ -82,79 +85,6 @@ where
     }
 }
 
-/// Set of [`Binary`] acting as a single [`Binary`]
-#[derive(Copy, Clone, Default, Debug)]
-pub struct Multi<C, B>
-where
-    C: BorrowMut<[B]>,
-    B: Binary,
-    B::Error: Miss,
-{
-    bins: C,
-    last: usize,
-    phantom: core::marker::PhantomData<B>,
-}
-
-impl<C, B> Multi<C, B>
-where
-    C: BorrowMut<[B]>,
-    B: Binary,
-    B::Error: Miss,
-{
-    /// Create a new [`Binary`] combining all `bins`
-    pub fn new(bins: C) -> Self {
-        Self {
-            bins,
-            last: 0,
-            phantom: Default::default(),
-        }
-    }
-}
-
-impl<C, B> From<C> for Multi<C, B>
-where
-    C: BorrowMut<[B]>,
-    B: Binary,
-    B::Error: Miss,
-{
-    fn from(bins: C) -> Self {
-        Self::new(bins)
-    }
-}
-
-impl<C, B> Binary for Multi<C, B>
-where
-    C: BorrowMut<[B]>,
-    B: Binary,
-    B::Error: Miss,
-{
-    type Error = B::Error;
-
-    fn get_insn(&mut self, address: u64) -> Result<Instruction, Self::Error> {
-        let bins = self.bins.borrow_mut();
-        let res = bins
-            .get_mut(self.last)
-            .map(|b| b.get_insn(address))
-            .filter(|r| !r.is_miss());
-        if let Some(res) = res {
-            return res;
-        }
-
-        let res = bins
-            .iter_mut()
-            .enumerate()
-            .filter(|(n, _)| *n != self.last)
-            .map(|(n, b)| (n, b.get_insn(address)))
-            .find(|(_, r)| !r.is_miss());
-        if let Some((current, res)) = res {
-            self.last = current;
-            res
-        } else {
-            Miss::miss(address)
-        }
-    }
-}
-
 /// [`Binary`] moved by a fixed offset
 ///
 /// Accesses will be mapped by subtracting the fixed offset from the address.
@@ -171,49 +101,5 @@ impl<B: Binary> Binary for Offset<B> {
 
     fn get_insn(&mut self, address: u64) -> Result<Instruction, Self::Error> {
         self.inner.get_insn(address.wrapping_sub(self.offset))
-    }
-}
-
-/// [`Binary`] adapter for an [`FnMut`]
-///
-/// This forwards calls to [`Binary::get_insn`] to the wrapped [`FnMut`].
-#[derive(Copy, Clone, Default, Debug)]
-pub struct Func<F: FnMut(u64) -> Result<Instruction, E>, E> {
-    func: F,
-    phantom: core::marker::PhantomData<E>,
-}
-
-impl<F: FnMut(u64) -> Result<Instruction, E>, E> Func<F, E> {
-    /// Create a new [`Binary`] from an [`FnMut`]
-    fn new(func: F) -> Self {
-        Self {
-            func,
-            phantom: Default::default(),
-        }
-    }
-}
-
-impl<F: FnMut(u64) -> Result<Instruction, E>, E> Binary for Func<F, E> {
-    type Error = E;
-
-    fn get_insn(&mut self, address: u64) -> Result<Instruction, Self::Error> {
-        (self.func)(address)
-    }
-}
-
-/// Create a [`Func`] [`Binary`] from an [`FnMut`]
-pub fn from_fn<F: FnMut(u64) -> Result<Instruction, E>, E>(func: F) -> Func<F, E> {
-    Func::new(func)
-}
-
-/// A [`Binary`] that does not contain any [`Instruction`]s
-#[derive(Copy, Clone, Default, Debug)]
-pub struct Empty;
-
-impl Binary for Empty {
-    type Error = error::NoInstruction;
-
-    fn get_insn(&mut self, _: u64) -> Result<Instruction, Self::Error> {
-        Err(error::NoInstruction)
     }
 }